test(autoload): fix PSR-4 test doubles
