+++ tests/Feature/Settings/PaymentGatewaySettingsTest.php
@@ -8,14 +8,8 @@ use App\Models\User;
 use App\Services\Payments\PaymentGatewayManager;
 use Database\Seeders\RoleAndPermissionSeeder;
 use Illuminate\Support\Collection;
-use OpenKOS\Core\Contracts\PaymentGateway;
-use OpenKOS\Core\Data\Payment\CheckoutInstructions;
-use OpenKOS\Core\Data\Payment\PaymentCreationResult;
-use OpenKOS\Core\Data\Payment\PaymentRequest;
-use OpenKOS\Core\Data\Payment\PaymentWebhookRequest;
-use OpenKOS\Core\Data\Payment\PaymentWebhookResult;
-use OpenKOS\Core\Enums\PaymentStatus;
 use OpenKOS\Platform\Payment\PaymentRegistry;
+use Tests\Support\Fakes\BillingTestPaymentGateway;
 
 uses()->beforeEach(function () {
     $this->seed(RoleAndPermissionSeeder::class);
@@ -183,57 +177,3 @@ it('forbids tenant-linked users from viewing and updating Billing settings', fun
         ])
         ->assertForbidden();
 });
-
-class BillingTestPaymentGateway implements PaymentGateway
-{
-    public function __construct(public array $config = []) {}
-
-    public function key(): string
-    {
-        return 'test/billing';
-    }
-
-    public function displayName(): string
-    {
-        return 'Billing Test Gateway';
-    }
-
-    public function createPayment(PaymentRequest $request): PaymentCreationResult
-    {
-        return new PaymentCreationResult(
-            providerReference: 'provider-reference',
-            status: PaymentStatus::Pending,
-            amount: $request->amount,
-            instructions: new CheckoutInstructions,
-        );
-    }
-
-    public function handleCallback(PaymentWebhookRequest $request): PaymentWebhookResult
-    {
-        return new PaymentWebhookResult(
-            eventReference: 'event-reference',
-            providerReference: 'provider-reference',
-            status: PaymentStatus::Pending,
-        );
-    }
-
-    public function configurationSchema(): array
-    {
-        return [
-            'environment' => [
-                'label' => 'Environment',
-                'type' => 'select',
-                'required' => true,
-                'options' => [
-                    ['value' => 'sandbox', 'label' => 'Sandbox'],
-                    ['value' => 'production', 'label' => 'Production'],
-                ],
-            ],
-            'secret_key' => [
-                'label' => 'Secret key',
-                'type' => 'password',
-                'required' => true,
-            ],
-        ];
-    }
-}
