+++ tests/Support/Fakes/MismatchedManagerTestPaymentGateway.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tests\Support\Fakes;
+
+class MismatchedManagerTestPaymentGateway extends ManagerTestPaymentGateway
+{
+    public function key(): string
+    {
+        return 'provider/gateway';
+    }
+}
