+++ tests/Unit/Services/PaymentGatewayManagerTest.php
@@ -2,15 +2,11 @@
 
 use App\Services\Payments\PaymentGatewayManager;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use OpenKOS\Core\Contracts\PaymentGateway;
-use OpenKOS\Core\Data\Payment\CheckoutInstructions;
-use OpenKOS\Core\Data\Payment\PaymentCreationResult;
-use OpenKOS\Core\Data\Payment\PaymentRequest;
-use OpenKOS\Core\Data\Payment\PaymentWebhookRequest;
-use OpenKOS\Core\Data\Payment\PaymentWebhookResult;
-use OpenKOS\Core\Enums\PaymentStatus;
 use OpenKOS\Platform\Payment\PaymentRegistry;
 use OpenKOS\Platform\Settings\SettingsManager;
+use Tests\Support\Fakes\BrokenManagerTestPaymentGateway;
+use Tests\Support\Fakes\ManagerTestPaymentGateway;
+use Tests\Support\Fakes\MismatchedManagerTestPaymentGateway;
 use Tests\TestCase;
 
 uses(TestCase::class, RefreshDatabase::class);
@@ -97,118 +93,3 @@ it('rejects gateways whose contract key differs from the registry key', function
     expect($gateway->find('registry/gateway'))->toBeNull()
         ->and($gateway->all()[0]['status'])->toBe('unavailable');
 });
-
-class ManagerTestPaymentGateway implements PaymentGateway
-{
-    public function __construct(public array $config = []) {}
-
-    public function key(): string
-    {
-        return 'test/gateway';
-    }
-
-    public function displayName(): string
-    {
-        return 'Test Gateway';
-    }
-
-    public function createPayment(PaymentRequest $request): PaymentCreationResult
-    {
-        return new PaymentCreationResult(
-            providerReference: 'provider-reference',
-            status: PaymentStatus::Pending,
-            amount: $request->amount,
-            instructions: new CheckoutInstructions,
-        );
-    }
-
-    public function handleCallback(PaymentWebhookRequest $request): PaymentWebhookResult
-    {
-        return new PaymentWebhookResult(
-            eventReference: 'event-reference',
-            providerReference: 'provider-reference',
-            status: PaymentStatus::Pending,
-        );
-    }
-
-    public function configurationSchema(): array
-    {
-        return [
-            'environment' => [
-                'label' => 'Environment',
-                'type' => 'select',
-                'required' => true,
-                'presentation' => 'segmented',
-                'default' => 'sandbox',
-                'options' => [
-                    ['value' => 'sandbox', 'label' => 'Sandbox'],
-                    ['value' => 'production', 'label' => 'Production'],
-                ],
-            ],
-            'webhook_setup' => [
-                'label' => 'Webhook setup',
-                'type' => 'info',
-                'instructions' => [
-                    'Open the webhook settings.',
-                    'Add the webhook URL shown below.',
-                ],
-                'link' => [
-                    'label' => 'Open webhook settings',
-                    'url' => 'https://example.test/webhooks',
-                ],
-                'url' => '/api/webhooks/test',
-            ],
-            'secret_key' => [
-                'label' => 'Secret key',
-                'type' => 'password',
-                'required' => true,
-                'description' => 'Keep this value secret.',
-                'visible_when' => [
-                    'field' => 'environment',
-                    'value' => 'sandbox',
-                ],
-            ],
-        ];
-    }
-}
-
-class BrokenManagerTestPaymentGateway implements PaymentGateway
-{
-    public function __construct()
-    {
-        throw new RuntimeException('broken gateway');
-    }
-
-    public function key(): string
-    {
-        return 'broken/gateway';
-    }
-
-    public function displayName(): string
-    {
-        return 'Broken Gateway';
-    }
-
-    public function createPayment(PaymentRequest $request): PaymentCreationResult
-    {
-        throw new RuntimeException('broken gateway');
-    }
-
-    public function handleCallback(PaymentWebhookRequest $request): PaymentWebhookResult
-    {
-        throw new RuntimeException('broken gateway');
-    }
-
-    public function configurationSchema(): array
-    {
-        return [];
-    }
-}
-
-class MismatchedManagerTestPaymentGateway extends ManagerTestPaymentGateway
-{
-    public function key(): string
-    {
-        return 'provider/gateway';
-    }
-}
