+++ tests/Unit/Services/WhatsAppManagerTest.php
@@ -8,13 +8,13 @@ use App\Models\Setting;
 use App\Services\WhatsAppManager;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\Event;
-use OpenKOS\Core\Contracts\WhatsAppDriver;
-use OpenKOS\Core\Data\WhatsApp\DriverHealthResult;
 use OpenKOS\Core\Data\WhatsApp\WhatsAppAttachment;
 use OpenKOS\Core\Data\WhatsApp\WhatsAppContent;
-use OpenKOS\Core\Data\WhatsApp\WhatsAppMessage;
 use OpenKOS\Platform\Notification\NotificationDriverRegistration;
 use OpenKOS\Platform\Notification\NotificationRegistry;
+use Tests\Support\Fakes\UnitTestFailingWhatsAppDriver;
+use Tests\Support\Fakes\UnitTestUnsupportedWhatsAppDriver;
+use Tests\Support\Fakes\UnitTestWhatsAppDriver;
 use Tests\TestCase;
 
 uses(TestCase::class, RefreshDatabase::class);
@@ -120,119 +120,3 @@ describe('WhatsAppManager', function () {
         });
     });
 });
-
-class UnitTestWhatsAppDriver implements WhatsAppDriver
-{
-    public static ?WhatsAppMessage $lastMessage = null;
-
-    public function __construct(public array $config = []) {}
-
-    public function send(WhatsAppMessage $message): void
-    {
-        self::$lastMessage = $message;
-    }
-
-    public function supportsAttachments(): bool
-    {
-        return true;
-    }
-
-    public function health(): DriverHealthResult
-    {
-        return new DriverHealthResult(true, 'Healthy');
-    }
-
-    public function supportsPairing(): bool
-    {
-        return false;
-    }
-
-    public function configurationSchema(): array
-    {
-        return [];
-    }
-
-    public function getPairingQrCode(): ?string
-    {
-        return null;
-    }
-
-    public function pair(): void {}
-
-    public function disconnect(): void {}
-}
-
-class UnitTestFailingWhatsAppDriver implements WhatsAppDriver
-{
-    public function __construct(public array $config = []) {}
-
-    public function send(WhatsAppMessage $message): void
-    {
-        throw new RuntimeException('Network connection failed');
-    }
-
-    public function supportsAttachments(): bool
-    {
-        return true;
-    }
-
-    public function health(): DriverHealthResult
-    {
-        return new DriverHealthResult(false, 'Unhealthy');
-    }
-
-    public function supportsPairing(): bool
-    {
-        return false;
-    }
-
-    public function configurationSchema(): array
-    {
-        return [];
-    }
-
-    public function getPairingQrCode(): ?string
-    {
-        return null;
-    }
-
-    public function pair(): void {}
-
-    public function disconnect(): void {}
-}
-
-class UnitTestUnsupportedWhatsAppDriver implements WhatsAppDriver
-{
-    public function __construct(public array $config = []) {}
-
-    public function send(WhatsAppMessage $message): void {}
-
-    public function supportsAttachments(): bool
-    {
-        return false;
-    }
-
-    public function health(): DriverHealthResult
-    {
-        return new DriverHealthResult(true);
-    }
-
-    public function supportsPairing(): bool
-    {
-        return false;
-    }
-
-    public function configurationSchema(): array
-    {
-        return [];
-    }
-
-    public function getPairingQrCode(): ?string
-    {
-        return null;
-    }
-
-    public function pair(): void {}
-
-    public function disconnect(): void {}
-}
